feat: remove server side load
