+++ package.json
@@ -76,9 +76,7 @@
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.51.3",
     "react-singleton-hook": "^3.4.0",
-    "remark": "^14.0.3",
     "rudder-sdk-js": "^2.48.6",
-    "strip-markdown": "^5.0.1",
     "styled-components": "^5.3.11",
     "typescript": "^5.4.5",
     "wagmi": "^2.12.17",
